refactor: Replaced dip, sip, fill_parent with dp, sp, match_parent
As well as formatting codes for clarity.
You may notice that I removed `<?xml version="1.0" encoding="utf-8"?>`, this is due to compile failure after switching to `dp` and `sp` because the version of the xml is 1 as the compiler complains, so I just let it manage the files itself.

Idk, why you guys didn't do this ages ago.